fix: deserialize based on schema fields
This fixes a bug where calls to Borsh's deserialize function were
silently failing to deserialize some fields. This is because the
constructor functions passed to deserialize expected camelCase keys
but the schemas are written with snake_case keys.

This fix generates the constructor parameter type from the associated
schema value, so constructors should be well-typed for the objects
passed in as arguments by Borsh.