fix: use es2015 in interface
This fixes a bug introduced in #317 where `new Query` cannot be called
because Query was transpiled to ES5 and relies on the class it extends
to also be transpiled, but the parent class was generated from Rust
and not transpiled.

PR #317 already fixed this for the extension; this commit just sets
the target JS version for namada-interface too.